hal: sdk-ng: sync sdk drivers to hal nxp
Added driver changelog to hal_nxp in this update.
The changelog documents what bugs this driver fixed,
what improvements were made, and what new features were
added during this upgrading.

+++ mcux/mcux-sdk-ng/drivers/syscon/fsl_syscon.c
@@ -1,7 +1,5 @@
 /*
- * Copyright  2018 NXP
- * All rights reserved.
- *
+ * Copyright 2018, 2025 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -36,7 +34,7 @@
  *
  * retval None.
  */
-void SYSCON_AttachSignal(SYSCON_Type *base, uint32_t index, syscon_connection_t connection)
+void SYSCON_AttachSignal(SYSCON_Type *base, uint16_t index, syscon_connection_t connection)
 {
     uint32_t syscon_id;
     uint32_t output_id;
